Add CreateDynamicSet<T> and CreateDynamicInstance<T>
<T> must implement IConvertible

If dynamic property cannot be cast to <T> then it remains a string.
Not sure if this is the behavior desired, or if an exception should
be thrown instead.

Tested for types: string, int, double, bool, and DateTime.

Motivation for this was to inhibit conversion logic when string was
the desired value type - more complicated scenarios may exist that
won't work.